+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Update `GetToolInstallationPath` to use user temp folder.
 - Update `TestRunUtil` icons to circle with color.
 - Update `AzureSignToolUtils` sign to ignore exception.
+- Update `AzureSignToolUtils` to ignore if file is already signed.
 ### Example
 - Add `Resource` and `Resource.pt-BR` to test sign files.
 ### Tests